util: ensure path-copy overwrites existing files
The utility method `pathCopy` uses distutil's `copy_tree` method to help
overwrite a destination directory's contents with the provided source
directory. The implementation originally (and incorrectly) set the
option `update=True` which results in existing files not being
overridden. This issue was first observed when using ``tox`` on a Linux
environment; it appears under some Windows Python implementations, the
`update` option was ignored.

Signed-off-by: James Knight <james.d.knight@live.com>