util: default to an error exit value when exit message is set
When invoking the `exit` method with an error message provided, the
default exit code set is a value of zero. In most cases where an exit
message is set, the result is an error case. Adjusting the default value
to be a value of one if a message is provided.

Signed-off-by: James Knight <james.d.knight@live.com>