util: allow path-remove to be quiet on error
When `pathRemove` results in an error attempting to remove the provided
path, an error message is generated. There may be cases where a caller
does not care if the path could not be removed even in the situation
where the removal fails due to a missing path entry. A caller can now
explicitly configure whether or not a failed call will output an error
message using the `quiet=True` argument.

Signed-off-by: James Knight <james.d.knight@live.com>