cvs: filter cvs folders using filter instead of exclude
Python's `TarFile` implementation defines an `exclude` argument (from
Python 2.7 to Python 3.7) as well as a filter argument (from Python 2.7
and onwards). Both can be used to achieve the desired results of
excluding the `CVS` directories when catching a CVS module; however,
releng-tool aims to support Python 3.7 and newer versions. Switching
the implementation to use the `filter` argument instead.

Signed-off-by: James Knight <james.d.knight@live.com>