support utility helper import
When releng-tool invokes a project's script (whether it be a
configuration script, a package stage-specific script (e.g. a package's
build script), etc.), a series of utility functions will be made
available with a pre-populated globals module. If a project defines
their own Python modules for script support, these modules will not have
a graceful way to take advantage of these utility methods. To assist
developers in this situation, utility functions will be included into
the `releng` namespace. A developer with a custom Python script can now
import specific utility methods using, for example, the following:

    from releng import releng_execute

Signed-off-by: James Knight <james.d.knight@live.com>