util: properly check for parent directory in file copy request
When attempting to copy a file using `pathCopy`, the request should
check if the target file's parent directory exists; if not, create it.
However, the check for the directory was checked on the target file
destination instead of the target file's parent directory (thus, would
always attempt to ensure the directory exists). Correcting by actually
checking for the parent directory before attempting to create it.

Signed-off-by: James Knight <james.d.knight@live.com>