util: ensure quiet copy request is quiet
When a caller invokes `pathCopy` with `quiet=True` argument, the
operation should not be outputting any errors to standard error if an
issue is detected. The internal call to `ensureDirectoryExists` may
result in an error being output. Correcting by also adding a quiet
option to `ensureDirectoryExists` and adjusting `pathCopy` to take
advantage of it.

Signed-off-by: James Knight <james.d.knight@live.com>