engine: re-work argument preparations
Various package types has the ability for a releng-tool package
developer to define options and environment variables for various
stages. In most cases, argument options built for the packages were done
so incorrectly or inconsistently. The adjustments ensure options are
provided to the execution request as proper argument entries.

The introduced changes prevent the ability for a CMake-package to
override CMake definitions. Definitions values bind their key-value
pairs as individual arguments, which need to be processed outside
configured options. A future commit will permit the ability for these
definitions to be defined and overridden by a package definition (if
desired).

Signed-off-by: James Knight <james.d.knight@live.com>