engine: introduce post-processing stage for individual packages
In certain build environments, a developer may wish to perform a series
of unique actions based on the completion of a specific package. To
allow this, a post-processing script `<pkg>-post>` can be defined in a
package's definition folder and will be invoked after the installation
stage of the respective package.

Signed-off-by: James Knight <james.d.knight@live.com>