cvs: force cvs-rsh to ssh
Tweak the `CVS_RSH` environment variable to always assume an SSH-based
authentication process when a developer configures an :ext: cvsroot.
This is most likely the authentication process (when using extensions)
used by anyone still needing to pull sources from CVS systems.

This can be re-worked in the future if developers have issues
authenticating with their CVS repositories with this fixed option.

Signed-off-by: James Knight <james.d.knight@live.com>