re-work forwarded arguments
The initial implementation of forwarding arguments from a releng-tool
set of used arguments to a project set of arguments was limited. A
project could only take advantage of dash-led arguments since any
first-word provided would assumed to be the releng-tool's target action.
To address this, pass arguments to a project configuration through the
use of the "--" argument. Leading values before this entry with be for
releng-tool use only; the remainder is left to the project
configuration.

Signed-off-by: James Knight <james.d.knight@live.com>