util: expand sequence to have an empty key value instead of none
When a zero-to-one-strings parsing request occurs, it has the ability to
convert a sequence (such as a list) into a series of dictionary entries.
Originally, entries would be added into the dictionary as keys with
values configured to `None`. This is not ideal when processing
configuration options as entries with values set to `None` are assumed
to be entries queued for removal (i.e. a developer wishes to signal not
to include an internally set configuration entry). Adjusting the default
value to an empty string to allow stages to know that the respective
options are desired, but just the value portion is desired to be ignore.

Signed-off-by: James Knight <james.d.knight@live.com>