git: remove progress from checkout request
When a Git project is processing the extraction phase, the checkout
request has a `--progress` argument to explicitly indicate to show the
progress of extraction; however, some common legacy systems do not have
this option available for use (i.e. CentOS 7). Removing the option as it
is not critical to this operation.

Signed-off-by: James Knight <james.d.knight@live.com>