opts: rename generic tool overrides to extraction tool overrides
Originally, the override option `override_tools` was added to allow a
user to override select tools inside the releng-tool process when
needed; however, the `override_tools` option was initially used and
documented incorrectly. `override_tools` is specifically used for
allowing a user decide an extraction tool type to use when translating
an extension type to a tool type (e.g. to support archive types not
included in releng-tool). To address this, the option will be renamed to
`override_extract_tools` to be more explicit in its purpose.

Signed-off-by: James Knight <james.d.knight@live.com>