engine: skip packages without license files during full build
When a user solely requests to build license information using the
`releng-tool licenses` call, the generated license file will contain
only packages that define license files to populate against. When
performing a standard build (which also will generate license
information), packages without license files will be added and have
empty sections in the generated license document. This is a result of a
package generating into the tracked license file dictionary when it has
no license files to populate with; correcting.

Signed-off-by: James Knight <james.d.knight@live.com>