Documentation: correct clean/mrproper descriptions
The original documentation described that the `clean` operation would
clean the entire output directory and the `mrproper` operation did that
along with cleaning cache/file-flags; however, this is no the case. The
clean operation only removes a subset of folders from inside the output
directory (to clean the build but not affect already built images and
file modes). The mrproper operation helps clean images and configured
file modes. None of the clean attempts should ever be attempting to
remove downloaded assets and cache folders. Correcting the documentation
to reflect on what releng-tool actually does.

Signed-off-by: James Knight <james.d.knight@live.com>