util: permit none-value for dictionary parsing
When parsing a dictionary for strings (`interpretDictionaryStrings`),
allow values of keys to be set to `None`. This will allow configuration
parsing to determine the difference when a package's configuration is
indicating between wanting to set an empty value for a key or remove the
use of the key (from internal defaults). This is same approach already
performed in `interpretZeroToOneStrings`.

Signed-off-by: James Knight <james.d.knight@live.com>