engine: handle package licenses in subdirectories
When processing configured license documents to extract from package
sources, license files are not guaranteed to be in the root of the
package. While license processing can read the documents, the operation
fails since the target subdirectory does not exist in the output license
folder. Adjusting the copy call from using `shutil` (directly) to the
utility `pathCopy` call to automatically deal with directory creation.

Signed-off-by: James Knight <james.d.knight@live.com>