introduce no-extract option for packages
In select cases, a developer may wish to fetch contents for a package
but not directly operate on the content. This can be for cases where
another package may deal with an archive of sources or a package archive
is intended for the final target.

While a developer specifying `LIBFOO_NO_EXTRACTION` on a package will
help skip the extraction phase, it will be the responsibility of the
developer to avoid using features which require extraction. For example,
if a developer also specifies license files to verify/manage, the
package processing will fail since it will have no extracted sources to
compare/manage on.

Signed-off-by: James Knight <james.d.knight@live.com>