main: handle forward arguments when dashes are the initial argument
releng-tool should detect a "--" argument to allow capturing a subset of
arguments to forward to engine options (for project-specific
customizations). The initial implementation would properly find the
index of an existing "--" argument; however, a first index will result
in a value of zero and would in turn fail the argument partitioning.
This commit allows a proper check for forwarded arguments by using "-1"
(instead of `None`) to check the validity of the found index (if any).

Signed-off-by: James Knight <james.d.knight@live.com>