engine: introduce bootstrapping stage for individual packages
In certain build environments, a developer may wish to perform a series
of unique actions before configuring a specific package no matter which
package type is defined. To allow this, a bootstrapping script
`<pkg>-bootstrap` can be defined in a package's definition folder and
will be invoked before the configuration stage of the respective
package.

Signed-off-by: James Knight <james.d.knight@live.com>