git: ensure hash-defined revision exists when fetching
When a fetch stage is processed for a Git repository, the configured
revision may be a full SHA-1 hash. In a scenario where a project's cache
directory is already populated without a configured hash value, a user
may experience the following checkout error during the extraction phase:

    fatal: reference is not a tree: <hash>
    (error) unable to checkout revision

The fetching implementation uses `rev-parse` command to help verify the
existence of various reference types. While the command's usage in this
scenario works for various reference types (branches, tags,
short-hashes, etc.), when a full hash is provided, the command will
successfully return for a valid hash value which may not exist in the
repository. To account for this, when a `rev-parse` request succeeds
and the output (the hash reference of the provided revision) matches the
revision, we know that the provided revision is a full SHA-1 hash. This
allows an additional sanity check where `cat-file` can be used to verify
that the full hash actually references a commit object.

Signed-off-by: James Knight <james.d.knight@live.com>