hg: fix broken fetch implementation
The implementation for Mercurial support was never published in a
working state resulting in fetch operations always failing with:

    NameError: name 'has_cache' is not defined

This was a result of some original cache consistently checking (via an
`hg verify` invoke) which took extremely too long and was cut.
Unfortunately, this removal did not properly replace the `has_cache`
variable with a proper cache-check; correcting.

Signed-off-by: James Knight <james.d.knight@live.com>