fetch: provide no-cache hint to fetchers
Adjusting the `RelengFetchOptions` operation to introducing a new
option `ignore_cache` to indicate whether or not a hint/flag has been
set to inform the fetching implementation that it should try to ignore
the existence of the cache (if possible). This is to help deal with
situations where a force-fetch may be desired even through a configured
revision already exists in a cached directory.

Signed-off-by: James Knight <james.d.knight@live.com>