fetch: allow git/mercurial calls to ignore cache
For development-based builds, a package may have a flag set to indicate
to ignore cached content (`LIBFOO_DEVMODE_IGNORE_CACHE`); however, for
DVCS-based packages, this flag did not help fetch new sources for
branch-like revision targets. A recent commit [1] introduced the ability
to notify a fetch operation that an "ignore cache" state is set -- this
change adjusts the Git and Mercurial implementations to force-fetch new
content even through the configured revision is already detected in the
respective package cache folders.

[1]: f9dcf86452f791be862aa3d64d2a5633f7ca1abd

Signed-off-by: James Knight <james.d.knight@live.com>