util: ensure bufsize is appropriately set
When releng-tool opens up another process, the original intent was to
perform line-buffering when possible to provide feedback to users in an
optimal way. There was also the introduction of polling support to help
deal with processes which cannot be gracefully monitored based off of
lines. Modifications eventually lead to setting `bufsize=1` to use
line buffering, but the option was invalid as processes where also
switched to open in binary mode.

This commit attempts to correct processing interaction by handling a
default and polling configurations. The default configuration will start
the processing in text mode and perform line buffering when reading
output (for prints or captures). Since the default mode is in text mode,
the read lines do not need to be decoded. For polling, the processed
will be unbuffered and releng-tool will read a single byte at a time.

Signed-off-by: James Knight <james.d.knight@live.com>