Documentation: restore landing page in toctree setup
Introducing release notes [1] also forced the toctree to be hidden to
prevent release note headings from being added in the document's main
page; however, this adjustment also removed the landing page from the
side-menu (not desired). Restoring the original toctree setup and append
the release notes document at the tail of the list but as "hidden".

[1]: cd117789c80b4173317af315cfc37fa5d5edea83

Signed-off-by: James Knight <james.d.knight@live.com>