util: handle removal of access-restricted paths
A package for a releng-tool project may create directories or files in
a way which prevent the cleanup stage from removing content. For
example, a package may install built files into the target directory as
read-only. This following introduces an internal way to deal with the
removal of a path (directories/files) which mimics the same way
`shutil.rmtree` works; however, adjusts the file mode on directories and
files to provide a better chance for a removal chance to succeed.

Signed-off-by: James Knight <james.d.knight@live.com>