util: permit moving directory over existing file
If a caller wishes to move a directory to a target location which is
already being used for a file, allow the operation to succeed by
replacing (removing) the file with the directory.

While may not typically be used for individual move requests, this
change is more helpful for larger moving requests where directory
content/structure may have significantly changed between package
revisions.

Signed-off-by: James Knight <james.d.knight@live.com>