fetch: sync-origin on git cached content
A Git-fetch operation will ensure that the repository's origin is
configured to the package's defined site. However, in the event that the
site changes and the cached content already holds the desired revision,
the origin is never updated to the new site. Updating the origin to the
configured site on a run helps ensure a user/developer manually
tinkering with the cache are using the expected site location without
having to manually adjust the origin as well.

This change also helps deal with an interim issue where the extraction
phase may attempt to pull additional content from a configured branch.
If a package is configured for a branch and has already been cached, an
origin update will pull updates (undesirable) during the extraction
phase (to be addressed in the future; developers should be using a
feature like `LIBFOO_DEVMODE_IGNORE_CACHE` instead); however, if the
origin is not synchronized, the extraction phase will fail with a remote
access error. This change prevents the extraction phase from blocking
(assuming the remote is available for access).

Signed-off-by: James Knight <james.d.knight@live.com>