fetch: git: reset on revision instead of pull
When the extraction phase for a Git site attempts to ensure the local
working tree is synchronized against the cached origin, it would perform
a `git pull` request for the branch; however, this request can result in
a network-fetch, which is not desired during the extraction phase.
Instead, replace the `pull` request with a `reset` request to ensure the
working directory is using the expected origin content.

If a user/developer was taking advantage of this "undesired" pull
operation, a better choice would be to investigate the use of
`LIBFOO_DEVMODE_IGNORE_CACHE` [1] instead.

[1]: https://docs.releng.io/en/latest/developer-guide.html#libfoo-devmode-ignore-cache

Signed-off-by: James Knight <james.d.knight@live.com>