Documentation: note that dvcs may not re-fetch content
The `LIBFOO_DEVMODE_IGNORE_CACHE` feature makes an indication that
"non-distributed version control [systems]... will not attempt to
re-fetch from a site since a cache file already exists"; however, this
is not true since DVCS fetch implementations, such as Git, may not
re-fetch as well. Adjusting the documentation to clear that if a
developer wishes to use dynamic content, it is advised to use the
ignore-cache capability (or just manually clean the cache directory
themselves).

Signed-off-by: James Knight <james.d.knight@live.com>