engine: pass quirk information to additional stages
Quirk information originally has only been passed into the engine's
build stage, as it was the only stage to use the single-supported quirk
option. To order to be be more flexible for other stages which may wish
to take advantage of quirk options in the future, pass these options
into the various stages as well.

Signed-off-by: James Knight <james.d.knight@live.com>