fetch: git: introduce default depth value
In most use cases, there is not a need for a releng-tool project to
fetch the entire history of a Git-based source into the local cache. The
following commit introduces the use of the `--depth` [1] argument when
fetching sources, to only fetch the tips of a Git-site's default/
configured references.

A developer does have the ability to adjust the depth value used (if
any), by explicitly configuring the package option `LIBFOO_GIT_DEPTH` to
the specific depth desired; where a depth value of `0` disables limited
fetching.

An exception exists in the event where a package defines a hash revision
which cannot be found. In this case, if the depth value is implicitly
set to the default of `1`, the fetch operation will instead fetch all
history of the repository (`--unshallow`). This is to be flexible for
simple configurations which may want to define a specific hash (that has
not been tagged yet) without having to deal with a depth option.

This commit also introduces a `releng.git.no_depth` quirk which allows a
caller to always prevents the use of the `--depth` argument from being
applied to any Git-based site.

[1]: https://git-scm.com/docs/git-fetch#Documentation/git-fetch.txt---depthltdepthgt

Signed-off-by: James Knight <james.d.knight@live.com>