log: flush stdout before flushing stderr
When an internal error message is generated, the error stream will be
flushed immediately (desired); however, if the error is a result of a
previous operation which has logged to stdout, logs can appear to be
out-of-order. To help limit this from occurring, stdout will be flushed
just before stderr is processed.

Signed-off-by: James Knight <james.d.knight@live.com>