packages: adding new local vcs-type
This commit introduces a new VCS-type defined `local`. This type
attempts to provide flexibility for new releng-tool projects which
select packages are not yet published to a site but wish to be tested
inside a releng-tool project. While "local sources mode" [1] provides a
way for users to work with internal packages with pre-fetched content,
package content needs to be prepared in specific relative parent
folders (which is not always ideal for development). The `local`
VCS-type allows a developer to place content/build files alongside a
package's definition file to rapidly start local development. For
example:

    ./package/my-test-package
        src/helloworld.cpp
        CMakeLists.txt
        my-test-package
            MY_TEST_PACKAGE_TYPE = 'cmake'
            MY_TEST_PACKAGE_VCS_TYPE = 'local'
            MY_TEST_PACKAGE_VERSION = '0.0.0'

Note that the introduction of the VCS-type does not intend to promote
the storing/committing of package/module's data alongside the definition
file. The sole reason why this VCS-type should be used is for
initial/interim development. To help promote this idea, a warning will
always be generated when a package uses this VCS-type.

[1]: https://docs.releng.io/en/latest/user-guide.html#local-sources-mode

Signed-off-by: James Knight <james.d.knight@live.com>