remove unused shebang
All of the internal scripts of the releng-tool project initially used
the same type of template header, which included an interpreter to use
(an environment-defined Python instance) if manually invoked; however,
none of these scripts are indented to be manually invoked. Cleaning up
the implementation by removing all the shebangs.

Signed-off-by: James Knight <james.d.knight@live.com>