support alternative script filenames
Various releng-tool scripts defined for a project have been (officially)
extension-less files. For example, a script-based package's
configuration script would be named the following for a `test-package`
package:

    test-package/test-package-configure

However, some developers may wish to define extensions for these files.
A previous commit [1] introduced some unofficial support for adding a
`releng` extension on package-specific files, but there is no specific
reason why a `releng` extension was chosen over a `py` extension
(releng-tool scripts are all Python-based scripts). It is assumed it
would be more ideal to stick with the common `py` extension.

This commit provides a more standard way to allow developers to define
various scripts with an extension. Specifically, any script documented
can be supported with and without a `.py` extension (by default, opting
for the extension-less variant). While a longer term goal (at this time)
is to support both ways of defining these files, alternative filenames
will not be promoted until the deprecated `releng` namespace is dropped
by this implementation [2] (due to a conflict with `releng.py` and the
deprecated namespace).

[1]: 39e6e66ef3e80a05cd6da8c569001952614d16b4
[2]: 3d5e9192e29e6fef38bb477cf4d891b654db53b4

Signed-off-by: James Knight <james.d.knight@live.com>