engine: always register optional flags in script env
A series of environment/script definitions are configured for a package
stage to assist in various implementation use cases. There exists a
subset of optional flags which are only set for their respective
conditional state, which nicely works for environment-based checks where
the flag can be checked without explicitly checking the contents of the
flag.

While this works for environment-based checks, script-based checks
require additional effort to check the existence of a flag. For example,
a script cannot explicitly check for a define's existence:

    if RELENG_RECONFIGURE:
        # todo

Without generating a undefined error in the script. Alternatively, a
developer would perform a safer check such as:

    if 'RELENG_RECONFIGURE' in os.environ:
        # todo

The former example would be preferred and can be achieved if the define
is set to a way which evaluated into a `False` condition by default.
Instead of always defining these conditional flags in both the
environment and script context's with a '1' or 'true' state, we can
gracefully support both context's by only defining the flags as `None`
inside the script context only. This allows the Python-based scripts to
check the flag options without worry of the define exists while still
allowing environment-based checks to check for the existence of the
define rather than examining the contents of the flag.

Signed-off-by: James Knight <james.d.knight@live.com>