allow configuring repository-specific git configurations
While a Git-defined site for a package provides the ability to prepare
a working directory for a configure revision, there is no specific way
a developer can influence internally managed Git actions for a specific
repository (outside adjusting a global configuration). For example, but
not limited to, configuring unique line-ending (`core.eol`), managing
symbolic links (`core.symlinks`) and applying only a spare-checkout
(`core.sparseCheckout`) of the repository.

A developer can now define a subset of unique configuration entries to
be applied after a bare repository has been created. For example:

    LIBFOO_GIT_CONFIG = {
       '<key>': '<value>',
    }

Signed-off-by: James Knight <james.d.knight@live.com>