util: correct execute's mixing of capture and quiet modes
When invoking `execute` (or `releng_execute`), the call includes a
`quiet` option to silent the output from stdout and a `capture` option
to capture the output into a list. The original intent of the execution
function would automatically silence the output to stdout when
capturing (unless the `quiet` option was explicitly set to `False`).
This commit corrects the implementation to what was intended.

Introducing a series of tests to help verify these expected states.

Signed-off-by: James Knight <james.d.knight@live.com>