ignore flake8 errors for import access points
releng-tool promotes callers import from the `releng_tool` module for
various definitions to include in their own implementations. flake8
makes the claim that they are not used; however, the alias imports are
documented for users to take advantage of.

This change also ignores the wildcard imports in the legacy namespace as
they we always expected a 1-to-1 mapping until the `releng` namespace is
dropped.

Signed-off-by: James Knight <james.d.knight@live.com>