scripts: removing obsolete git attributes
Removing an obsolete `.gitattributes` file previously used for
completion scripts line endings configuration. The original bash
completion script (`releng-tool-completion`) was moved during a cleanup
commit [1]. The attributes file was never updated, but was addressed in
a previous commit [2] with a different attributes file. Since this file
is incorrect and no longer needed, removing.

[1]: 7f52360b29fa2e23f84664c9251094679dd56427
[2]: d8d0ae8ab399c77811df0ab519c7a65acffc2b24

Signed-off-by: James Knight <james.d.knight@live.com>