prerequisites: introduction
Adding support to check for prerequisites for a releng-tool process,
which is a nice way to say "checking for host tools before running".
This commit introduces two types of prerequisite checking.

First, an internally managed prerequisite list will be generated based
on the configured packages. For example, if a package clones from a Git
repository, the existence of `git` will be checked before running any
package stages. A series of internal checks have been added for package
types and version control system types.

Second, users can specify a list of host tools they want to check for
using a new `prerequisites` project configuration key. This contains a
list of strings which releng-tool which attempt to find on the host
system before running package stages. For example, a developer may
specify the need for `makepkg` which is used in a project's
post-processing script. releng-tool which check for the tool's existence
on the system instead of building each package and then failing in the
post-processing stage.

Signed-off-by: James Knight <james.d.knight@live.com>