prerequisites: introduction
The introduction [1] of a prerequisites check would general an error
message for missing host tools; however, the initial implementation
would print an incomplete error message when all required host tools (if
any) are found. This commit corrects the error state by only attempting
to generate an error message when one or more host tools are missing.

[1]: 6504f48b718f5678b4ebc7d1b5178b62d275f8cb

Signed-off-by: James Knight <james.d.knight@live.com>