introduce rebuild-only/reconfigure-only actions
The initial implementation for a package-specific `rebuild` or
`reconfigure` actions would re-perform the respective stage and stop. In
most scenarios, a user wanting to perform, for example, a reconfigure of
a package would want to re-do the configuring and continue with the
remaining stages of that specific package (i.e. building and installing
stages). A user can achieve this by invoking the `<pkg>-reconfigure`
action, followed by the `<pkg>` action; but this is not ideal.

Instead, the `rebuild` and `reconfigure` actions will re-do their
respective stage and complete the remaining stages of the package. In
the case where a developer does not want to perform anything but the
specific stage, `-only` variants of these actions have been created. To
only invoke the force just the reconfiguration of a package, a developer
will now invoke `<pkg>-reconfigure-only`. Likewise for only invoking the
rebuilding of a package will now be done via `<pkg>-rebuild-only`.

Signed-off-by: James Knight <james.d.knight@live.com>