set script/environment flags when in a clean action
Project scripting may wish to take advantage of scenarios where a user
invokes a clean action (clean, mrproper or distclean), but there is not
a proper way to determine if a clean action has been set. The following
commits adds the `RELENG_CLEAN`, `RELENG_DISTCLEAN` and
`RELENG_MRPROPER` variables to the environment (for the respective
actions).

Note that the level of cleaning is dictated by the requested action,
with `clean` being the "lightest" and `distclean` being the "strongest".
When a stronger clean action is performed, all the ligher variables will
be set as well. For example, if `mrproper` action is invoked, both
`RELENG_CLEAN` and `RELENG_MRPROPER` variables will be set.

Signed-off-by: James Knight <james.d.knight@live.com>