util: verbose dump execute arguments in an easier format
When in a verbose-enabled mode, executed commands are outputted to the
configured standard output stream. However, the printed output would
dump the commands in list format which is not ideal for users trying to
debug commands (i.e. trying to copy-paste commands to invoke them out of
the releng-tool process).

Instead, commands will be formatted into a style which better suits the
active environment -- `subprocess.list2cmdline` is used to better format
for `cmd.exe` environments, and `[shlex|pipes].quote` to format for all
other environments.

Signed-off-by: James Knight <james.d.knight@live.com>