util: debug dump working directory for execute calls
When debugging a project, a developer may enable the verbose option to
dump commands invoked on the command line; however, there is no
indicatation to which working directory the command is being invoked
under. This commit provides users the ability to dump the working
directory of invoked commands when the debug option is set.

Signed-off-by: James Knight <james.d.knight@live.com>