engine: rework a subset of error states to raise exceptions
The releng-tool engine primarily handled error scenarios by dumping a
message to the configured standard error string and returned from the
running method. While this works for a user invoking the releng-tool, it
is not trivial to attempt to test/capture error scenarios in unit
testing.

The following adjusts two error scenarios from an error message to a
raised exception, where (1) missing configuration will now raise a
`MissingConfigurationError` exception and (2) if no packages are defined
in a configuration, a `MissingPackagesError` exception will be raised.
This will allow future unit tests to check for these error scenarios.

The releng-tool mainline has been adjusted to always capture releng-tool
generated messages and print them to the configured standard error
stream, so users should experience the same style of error events.

Signed-off-by: James Knight <james.d.knight@live.com>