only apply default sysroot prefix option when building stage options
Instead of assigning a package's target prefix value to a default
sysroot prefix (from the project's configured options) when building a
package instance, only hint to use the sysroot prefix when building a
stage's options. This allows the engine to determine whether or not the
package has been explicit configured with a prefix by checking for a
`None` assignment; which is to help in future development where a
package's prefix value can be late-stage loaded into the package entity.

Signed-off-by: James Knight <james.d.knight@live.com>