packages: full stop if external package definition fails to load
Package processing will cycle through a list of configured external
package folders before attempting to load the project's "main" package
folder. In the event that an external package is found and fails to
load, this can result in the package loading to unexpectedly falling
back on another package definition (if defined). The releng-tool process
may attempt to build with the subsequent package definition and a user
would only be able to tell if they manually inspect the logs.

Instead, if a package folder defines a package definition to load
against, only attempt to load (and fail hard) if there are issues
attempting to load the configured package target.

Signed-off-by: James Knight <james.d.knight@live.com>