packages: directly configure autotools autoreconf flag
Directly attempts to apply a possible autotools autoreconf flag value,
which allows the releng-tool engine to determine whether or not the
flag has been explicitly set (`True` or `False`) or implicitly set (a
value of `None` which defaults to `False` evaluation when used). This is
to prepare for prospect late-stage configuration loading for a package.

Signed-off-by: James Knight <james.d.knight@live.com>