packages: raise package configuration error on invalid install type
Adjusts the package configuration processing to raise an
`RelengToolInvalidPackageConfiguration` exception when an invalid
install type is detected, instead of printing an error message and
returning a "no package" return value. This will help callers to handle
more specific errors (currently, to be used for upcoming unit test
work).

This change should result in the same experience for users as all
`RelengToolException`-based exceptions will be converted into an error
message before exiting.

Signed-off-by: James Knight <james.d.knight@live.com>