log: do not format if no arguments are applied
While the logging framework provides a helper means to format log
entries provided by optional arguments, logging may break if a provided
message/format is provided which includes contents that the format
operation will throw an exception.

To avoid this, if no arguments are provided, do not attempt to format
arguments into the message. For cases where users define arguments for
formatting, it is the callers responsibility to ensure the format
message is well formed for a format request.

Signed-off-by: James Knight <james.d.knight@live.com>