support loading remote package configuration
Typically, package definitions in a project will define all the package
configuration options in each package's "definition file". For
environments which reuse the same package over different releng-tool
projects and do not take advantage of external packages, a developer may
wish to shift the responsibility of these configurations into the
package itself.

With this commit, when a package has been fetched/prepared, if the
source of the package contains a `.releng-tool` file, the file will be
used to load supported package configurations that have not been
explicitly configured yet. Options which are set in the releng-tool
project take precedence over options defined on the package's contents.
Developers also have the ability to force ignore remote configurations
through the use of a `LIBFOO_SKIP_REMOTE_CONFIG` flag.

Signed-off-by: James Knight <james.d.knight@live.com>