support loading remote package scripts
Typically, script-based package in a project will provide scripts in the
package's definition folder. For environments which reuse the same
package over different releng-tool projects and do not take advantage of
external packages, a developer may wish to shift the responsibility of
these scripts into the package itself.

With this commit, when a script-based package invokes a stage script, if
the package does not provide a script for the stage, the script will
also be looked for in the root of the package's contents. Developers
have the ability to force ignore remote scripts through the use of a
`LIBFOO_SKIP_REMOTE_SCRIPTS` flag.

Signed-off-by: James Knight <james.d.knight@live.com>