util: improve error message for missing path_copy src
When requesting to copy a file/directory via the `path_copy` utility
call, the logic assumes its a directory if a file type is not detected.
The error in the case where `src` does not exist may not be clear to
users if the intended source was a file (i.e. the error message will
indicate the directory does not exist instead of a file). This commit
adjusts the error case to report a specific error message if the
provided source path does not exist.

Signed-off-by: James Knight <james.d.knight@live.com>