tests: skip test_utilio_execution is verbose
The `test_utilio_execution` unit test operates on invoking and capturing
specific information on the standard output/error streams. Enabling
verbosity on a unit test will cause the unit test to fail. In the event
that a verbose mode is detected while running this test, the specific
unit test will be skipped.

Signed-off-by: James Knight <james.d.knight@live.com>